src/requirements.txt
discord
python-dotenv
replicate
openai